Remove (series::install)
The PERIODS package only needs to refer to symbols in the SERIES
packages, which is already achieved by the :use clause (see package
definition). If the install step was really necessary at compilation
time (so as to rebind LET, ...), it should have been wrapped in an
appropriate EVAL-WHEN form, which was not the case.